refactor out common logic
